Some router re-jigging
