+++ apps/website/src/server/routers/facilitators.ts
@@ -1,0 +1,337 @@
+import {
+  and,
+  arrayContains,
+  courseFeedbackTable,
+  eq,
+  facilitatorDiscussionSwitchingTable,
+  groupDiscussionTable,
+  groupTable,
+  inArray,
+  meetPersonTable,
+  peerFeedbackTable,
+  roundTable,
+  unitTable,
+} from '@bluedot/db';
+import { TRPCError } from '@trpc/server';
+import z from 'zod';
+import db from '../../lib/api/db';
+import { protectedProcedure, router } from '../trpc';
+
+const getFacilitator = async (roundId: string, facilitatorEmail: string) => {
+  const facilitator = await db.getFirst(meetPersonTable, {
+    filter: { round: roundId, email: facilitatorEmail, role: 'Facilitator' },
+  });
+  if (!facilitator) {
+    throw new TRPCError({ code: 'NOT_FOUND', message: 'No facilitator found for this round' });
+  }
+  return facilitator;
+};
+
+async function verifyFacilitatorById(meetPersonId: string, email: string) {
+  const meetPerson = await db.getFirst(meetPersonTable, {
+    filter: { id: meetPersonId, email },
+  });
+  if (!meetPerson || meetPerson.role !== 'Facilitator') {
+    throw new TRPCError({ code: 'FORBIDDEN', message: 'Not authorized' });
+  }
+  return meetPerson;
+}
+
+async function getGroupForFacilitator(facilitatorId: string) {
+  const groups = await db.pg
+    .select({ id: groupTable.pg.id, groupName: groupTable.pg.groupName, participants: groupTable.pg.participants })
+    .from(groupTable.pg)
+    .where(arrayContains(groupTable.pg.facilitator, [facilitatorId]));
+  if (groups.length === 0) {
+    throw new TRPCError({ code: 'NOT_FOUND', message: 'No group found for this facilitator' });
+  }
+  return groups[0]!;
+}
+
+async function getPeerFeedbackForParticipants(participantIds: string[]) {
+  if (participantIds.length === 0) return [];
+  const all = await db.pg.select().from(peerFeedbackTable.pg);
+  return all.filter((pf) => pf.feedbackRecipient?.some((r) => participantIds.includes(r)));
+}
+
+export const facilitatorRouter = router({
+  // --- Switching ---
+
+  getFacilitatorsForRound: protectedProcedure.input(z.object({ roundId: z.string() })).query(async ({ input, ctx }) => {
+    const { roundId } = input;
+    const currentFacilitator = await getFacilitator(roundId, ctx.auth.email);
+
+    const facilitators = await db.pg
+      .select({ id: meetPersonTable.pg.id, name: meetPersonTable.pg.name })
+      .from(meetPersonTable.pg)
+      .where(and(eq(meetPersonTable.pg.round, roundId), eq(meetPersonTable.pg.role, 'Facilitator')));
+
+    return facilitators
+      .filter((f) => f.id !== currentFacilitator.id)
+      .map((f) => ({ value: f.id, label: f.name ?? '' }));
+  }),
+
+  discussionsAvailable: protectedProcedure
+    .input(z.object({ roundId: z.string() }))
+    .query(async ({ input: { roundId }, ctx }) => {
+      const facilitator = await getFacilitator(roundId, ctx.auth.email);
+
+      const groupDiscussions = await db.pg
+        .select()
+        .from(groupDiscussionTable.pg)
+        // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
+        .where(and(inArray(groupDiscussionTable.pg.id, facilitator.expectedDiscussionsFacilitator || [])));
+
+      const groups = await db.pg
+        .select()
+        .from(groupTable.pg)
+        .where(inArray(groupTable.pg.id, groupDiscussions.map((discussion) => discussion.group)));
+
+      const unitIds = [...new Set(groupDiscussions.map((d) => d.courseBuilderUnitRecordId).filter(Boolean))] as string[];
+      const units = unitIds.length > 0
+        ? await db.scan(unitTable, { OR: unitIds.map((id) => ({ id, unitStatus: 'Active' as const })) })
+        : [];
+
+      const groupMap = new Map(groups.map((g) => [g.id, g]));
+      const unitMap = new Map(units.map((u) => [u.id, u]));
+
+      return groupDiscussions.map((discussion) => {
+        const group = groupMap.get(discussion.group);
+        if (!group) {
+          throw new TRPCError({ code: 'NOT_FOUND', message: `Related group not found for discussion ${discussion.id}` });
+        }
+        return { ...discussion, groupDetails: group, unitRecord: unitMap.get(discussion.courseBuilderUnitRecordId ?? '') ?? null };
+      });
+    }),
+
+  updateDiscussion: protectedProcedure
+    .input(z.object({
+      roundId: z.string(),
+      discussionId: z.string().optional(),
+      groupId: z.string(),
+      requestedDateTimeInSeconds: z.number(),
+    }))
+    .mutation(async ({ input, ctx }) => {
+      const { roundId, discussionId, groupId, requestedDateTimeInSeconds } = input;
+
+      const nowInSeconds = Math.floor(Date.now() / 1000);
+      if (requestedDateTimeInSeconds <= nowInSeconds) {
+        throw new TRPCError({ code: 'BAD_REQUEST', message: 'Requested time must be in the future' });
+      }
+
+      const facilitator = await getFacilitator(roundId, ctx.auth.email);
+      // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
+      const allowedDiscussions = facilitator.expectedDiscussionsFacilitator || [];
+
+      if (allowedDiscussions.length === 0) {
+        throw new TRPCError({ code: 'FORBIDDEN', message: 'Facilitator is not allowed to manage any discussions' });
+      }
+
+      if (discussionId) {
+        if (!allowedDiscussions.includes(discussionId)) {
+          throw new TRPCError({ code: 'FORBIDDEN', message: 'Facilitator is not allowed to manage this discussion' });
+        }
+        const discussion = await db.getFirst(groupDiscussionTable, { filter: { id: discussionId } });
+        if (!discussion || discussion.group !== groupId) {
+          throw new TRPCError({ code: 'BAD_REQUEST', message: 'Discussion does not belong to the specified group' });
+        }
+      }
+
+      if (!discussionId) {
+        const groupDiscussions = await db.pg
+          .select({ id: groupDiscussionTable.pg.id })
+          .from(groupDiscussionTable.pg)
+          .where(and(eq(groupDiscussionTable.pg.group, groupId), inArray(groupDiscussionTable.pg.id, allowedDiscussions)));
+        if (groupDiscussions.length === 0) {
+          throw new TRPCError({ code: 'FORBIDDEN', message: 'Facilitator is not allowed to manage any discussions in this group' });
+        }
+      }
+
+      await db.insert(facilitatorDiscussionSwitchingTable, {
+        // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
+        discussion: discussionId || null,
+        facilitator: facilitator.id,
+        group: groupId,
+        status: 'Requested',
+        switchType: discussionId ? 'Change for one unit' : 'Change permanently',
+        facilitatorRequestedDatetime: requestedDateTimeInSeconds,
+      });
+
+      return null;
+    }),
+
+  requestFacilitatorChange: protectedProcedure
+    .input(z.object({
+      roundId: z.string(),
+      discussionId: z.string(),
+      groupId: z.string(),
+      newFacilitatorId: z.string(),
+    }))
+    .mutation(async ({ input, ctx }) => {
+      const { roundId, discussionId, groupId, newFacilitatorId } = input;
+
+      const facilitator = await getFacilitator(roundId, ctx.auth.email);
+      // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
+      const allowedDiscussions = facilitator.expectedDiscussionsFacilitator || [];
+
+      if (allowedDiscussions.length === 0 || !allowedDiscussions.includes(discussionId)) {
+        throw new TRPCError({ code: 'FORBIDDEN', message: 'Facilitator is not allowed to manage this discussion' });
+      }
+
+      const discussion = await db.getFirst(groupDiscussionTable, { filter: { id: discussionId } });
+      if (!discussion || discussion.group !== groupId) {
+        throw new TRPCError({ code: 'BAD_REQUEST', message: 'Discussion does not belong to the specified group' });
+      }
+
+      const nowInSeconds = Math.floor(Date.now() / 1000);
+      if (discussion.startDateTime <= nowInSeconds) {
+        throw new TRPCError({ code: 'BAD_REQUEST', message: 'Cannot change facilitator for a discussion that has already started' });
+      }
+
+      const newFacilitator = await db.getFirst(meetPersonTable, { filter: { id: newFacilitatorId } });
+      if (!newFacilitator) {
+        throw new TRPCError({ code: 'NOT_FOUND', message: 'New facilitator not found' });
+      }
+      if (newFacilitator.round !== facilitator.round) {
+        throw new TRPCError({ code: 'BAD_REQUEST', message: 'New facilitator must be in the same round' });
+      }
+      if (newFacilitator.role !== 'Facilitator') {
+        throw new TRPCError({ code: 'BAD_REQUEST', message: 'Selected person is not a facilitator' });
+      }
+
+      await db.insert(facilitatorDiscussionSwitchingTable, {
+        discussion: discussionId,
+        facilitator: newFacilitator.id,
+        group: groupId,
+        status: 'Requested',
+        switchType: 'Update discussion facilitator',
+        anythingElse: `Requested by facilitator ${facilitator.id}`,
+      });
+
+      return null;
+    }),
+
+  // --- Feedback ---
+
+  getFeedbackFormData: protectedProcedure
+    .input(z.object({ meetPersonId: z.string().min(1) }))
+    .query(async ({ input, ctx }) => {
+      const meetPerson = await verifyFacilitatorById(input.meetPersonId, ctx.auth.email);
+      const group = await getGroupForFacilitator(meetPerson.id);
+      const participantIds = group.participants ?? [];
+
+      const participants = participantIds.length > 0
+        ? await db.pg.select({ id: meetPersonTable.pg.id, name: meetPersonTable.pg.name }).from(meetPersonTable.pg).where(inArray(meetPersonTable.pg.id, participantIds))
+        : [];
+
+      const round = meetPerson.round
+        ? await db.getFirst(roundTable, { filter: { id: meetPerson.round }, sortBy: 'id' })
+        : null;
+
+      const existingCourseFeedback = (meetPerson.courseFeedback ?? []).length > 0
+        ? await db.getFirst(courseFeedbackTable, { filter: { id: meetPerson.courseFeedback![0]! }, sortBy: 'id' })
+        : null;
+
+      const existingPeerFeedback = await getPeerFeedbackForParticipants(participantIds);
+
+      return {
+        meetPersonId: meetPerson.id,
+        roundName: round?.title ?? '',
+        groupId: group.id,
+        groupName: group.groupName,
+        participants: participants.map((p) => ({ id: p.id, name: p.name ?? '' })),
+        existingCourseFeedback: existingCourseFeedback ? {
+          id: existingCourseFeedback.id,
+          courseRating: existingCourseFeedback.courseRating,
+          courseValue: existingCourseFeedback.courseValue,
+          improvements: existingCourseFeedback.improvements,
+          completed: existingCourseFeedback.completed,
+        } : null,
+        existingPeerFeedback: existingPeerFeedback.map((pf) => ({
+          id: pf.id,
+          recipientId: pf.feedbackRecipient?.[0] ?? '',
+          initiativeRating: pf.initiativeRating,
+          reasoningQualityRating: pf.reasoningQualityRating,
+          feedback: pf.feedback,
+          nextSteps: pf.nextSteps,
+        })),
+      };
+    }),
+
+  savePeerFeedback: protectedProcedure
+    .input(z.object({
+      meetPersonId: z.string().min(1),
+      participantId: z.string().min(1),
+      peerFeedbackId: z.string().optional(),
+      initiativeRating: z.number().min(1).max(5),
+      reasoningQualityRating: z.number().min(1).max(5),
+      feedback: z.string(),
+      nextSteps: z.array(z.string()),
+    }))
+    .mutation(async ({ input, ctx }) => {
+      await verifyFacilitatorById(input.meetPersonId, ctx.auth.email);
+
+      const fields = {
+        feedbackRecipient: [input.participantId],
+        initiativeRating: input.initiativeRating,
+        reasoningQualityRating: input.reasoningQualityRating,
+        feedback: input.feedback,
+        nextSteps: input.nextSteps,
+      };
+
+      if (input.peerFeedbackId) {
+        return db.update(peerFeedbackTable, { id: input.peerFeedbackId, ...fields });
+      }
+      return db.insert(peerFeedbackTable, fields);
+    }),
+
+  submitFeedback: protectedProcedure
+    .input(z.object({
+      meetPersonId: z.string().min(1),
+      courseRating: z.number().min(1).max(5),
+      courseValue: z.string(),
+      improvements: z.string(),
+      courseFeedbackId: z.string().optional(),
+    }))
+    .mutation(async ({ input, ctx }) => {
+      const meetPerson = await verifyFacilitatorById(input.meetPersonId, ctx.auth.email);
+      const group = await getGroupForFacilitator(meetPerson.id);
+      const peerFeedback = await getPeerFeedbackForParticipants(group.participants ?? []);
+      const peerFeedbackIds = peerFeedback.map((pf) => pf.id);
+
+      const courseFields = {
+        courseRating: input.courseRating,
+        courseValue: input.courseValue,
+        improvements: input.improvements,
+        completed: true as const,
+        personFeedback: peerFeedbackIds,
+      };
+
+      const courseFeedback = input.courseFeedbackId
+        ? await db.update(courseFeedbackTable, { id: input.courseFeedbackId, ...courseFields })
+        : await db.insert(courseFeedbackTable, {
+          person: [meetPerson.id],
+          round: meetPerson.round ? [meetPerson.round] : [],
+          ...courseFields,
+        });
+
+      await db.update(meetPersonTable, { id: meetPerson.id, courseFeedback: [courseFeedback.id] });
+
+      for (const pfId of peerFeedbackIds) {
+        await db.update(peerFeedbackTable, { id: pfId, courseFeedback: [courseFeedback.id] });
+      }
+
+      return { courseFeedbackId: courseFeedback.id };
+    }),
+
+  unsubmitFeedback: protectedProcedure
+    .input(z.object({ meetPersonId: z.string().min(1) }))
+    .mutation(async ({ input, ctx }) => {
+      const meetPerson = await verifyFacilitatorById(input.meetPersonId, ctx.auth.email);
+      await db.update(meetPersonTable, { id: meetPerson.id, courseFeedback: [] });
+      if (meetPerson.courseFeedback?.[0]) {
+        await db.update(courseFeedbackTable, { id: meetPerson.courseFeedback[0], completed: false });
+      }
+      return { success: true };
+    }),
+});
